requirements.txt
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
fpdf2==2.7.5
python-dotenv==1.0.0
